SCHEDULING-433 Changed type of parameters from IntWrapper to int in some method signatures of RMUser and RMCore.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@13963 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/resource-manager/tests/nodestate/GetAllNodes.java
@@ -11,7 +11,7 @@ public class GetAllNodes {
         try {
             auth = RMConnection.join(null);
             RMUser admin = auth.logAsUser("demo", "demo");
-            admin.getAtMostNodes(admin.getFreeNodesNumber(), null);
+            admin.getAtMostNodes(admin.getFreeNodesNumber().intValue(), null);
         } catch (Exception e) {
             e.printStackTrace();
         }
